requirements_dev.txt
APScheduler==3.6.3
bcrypt==3.2.0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
cryptography==36.0.1
paramiko==2.9.2
pycparser==2.21
PyNaCl==1.5.0
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tornado==6.1
tzdata==2021.5
tzlocal==4.1
